Update pong_training_right.py
The training works now with minor issue(s). Sometimes a player that should be eliminated from the population does not actually get removed. I think the issue is when the ball's edge attains a value of exactly 780 pixels. The solution will require a better way to detect collisions, and keep track of which player has collided and which hasn't.